+++ crates/filter-list-manager/src/manager/models/disabled_rules_raw.rs
@@ -1,0 +1,7 @@
+use crate::FilterId;
+
+/// List of disabled rules separated by line breaks + filter id
+pub struct DisabledRulesRaw {
+    pub filter_id: FilterId,
+    pub text: String,
+}