+++ crates/filter-list-manager/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.5] - 2024-11-05
+
+### Fixed
+- Disabled rules of filters were removed after filters update
+
+### Added
+- `get_disabled_rules` method
+
+[0.8.5]: https://github.com/AdguardTeam/FilterListManager/compare/flm-0.8.4...flm-0.8.5
+
 ## [0.8.4] - 2024-11-01
 
 ### Fixed